complete transfer_money api work

+++ src/transfer/controllers/delete.js
@@ -1,15 +1,23 @@
 import Response from '../../../class/response.js';
+import transfer_model from '../models/index.js';
 import deleteData from '../services/delete.js';
-import deCodeVerifiedToken from '../../../utils/playWithToken.js';
 
 const deleteController = async (req, res) => {
     const response = new Response(res);
 
+
     const { id } = req.params;
 
     try {
+
+        const isExists = await transfer_model.findById(id);
+        if (!isExists) {
+            return response.error({}, "No document found to delete.");
+        }
+
+
         const data = await deleteData(id);
-        return response.success({ key: 'value' }, 'Data fetched successfully');
+        return response.success({ user: id }, 'Data fetched successfully');
     } catch (err) {
         response.error({}, 'Failed to fetch data');
     }
